requirements.txt
Django==4.2.7
Pillow==10.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-jazzmin==2.6.0
psycopg==3.1.14
python-dotenv==1.0.0
django-solo==2.1.0
django-crum==0.7.9
django-ckeditor==6.7.0
django-extensions==3.2.3
django-import-export==3.2.0
psycopg2-binary==2.9.9



